implemented clock system

+++ src/rogue/systems/base_sys.cs
@@ -4,6 +4,6 @@ namespace Rogue.Systems
 {
     public abstract class BaseSys
     {
-        public abstract void Update(uint? delta);
+        public abstract void Update(float? delta);
     }
 }
